requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
customtkinter>=5.2.0
Pillow>=10.0.0
PyInstaller>=6.0.0
python-dateutil>=2.8.0
requests>=2.31.0